Dodano wszystkie możliwości odbicia wiązek lasera; lustra odbijają wiązki

+++ src/wjpgra/Laser.java
@@ -4,11 +4,11 @@ package wjpgra;
 import javafx.scene.shape.Line;
 
 public class Laser extends Line {
-    public int xs;
-    public int ys;
-    public int xk;
-    public int yk;
-    public int a;
-    public int b;
+    public double xs;
+    public double ys;
+    public double xk;
+    public double yk;
+    public double a;
+    public double b;
     
 }
